wildmatch: teach Basename to match against basename only
In Git's attr.c, patterns found in a repository's .gitattributes
file(s) are matched on basename alone if and only if the pattern
contains no directory separators [1], [2].

This behavior allows a line such as:

    *.txt filter=lfs -text

in the repository's root .gitattributes (or .git/info/attributes) to
apply the filter=lfs and text=false attributes to a.txt, and
/path/to/a.txt. In general terms, this treats patterns as having a
prefixed `**/` in this case.

To simulate this behavior, we introduce the wildmatch.Basename
option to truncate incoming paths to their basename (using
filepath.Base from package path/filepath) if and only if the given
pattern contains only 1 ``split'', i.e., does not contain directory
separators.

For instance, given the wildmatch.Basename option, the following
patterns match:

  - *.txt matches foo.txt
  - *.txt matches foo/bar.txt
  - foo/*.txt matches foo/bar.txt
  - foo/*.txt does not match baz/foo/bar.txt
  - foo/*.txt does not match foo/bar/baz/woot.txt

This option suits a tree-like representation of a repository's
.gitattribute file(s) nicely, by giving wildmatch.Basename, we mirror
the correct behavior for paths relative to the directory in which they
are expressed in a .gitattributes file.

[1]: https://github.com/git/git/blob/v2.18.0/attr.c#L1005-L1010
[2]: https://github.com/git/git/blob/v2.18.0/dir.c#L608-L613