wildmatch: allow escaping hash mark
Wildmatch patterns are often used in configuration files where the hash
mark is a comment character.  Currently, an attempt to escape this
character in order to avoid its interpretation as the comment character
causes the escape character to be treated as a path separator, which is
definitely not what the user expects.

To ensure wildmatch can be used in these kinds of configuration files,
allow escaping the hash mark.