Fix swapped case-sensitivity defaults
We have two implementations of the init function, which sets the system
case sensitivity: one for Linux, and one for non-Linux.  However, the
implementations are swapped: the Linux one is case insensitive, and the
non-Linux is case sensitive.  Correct this by swapping the functions
around so that Linux is case sensitive and Windows and Mac are case
insensitive.