Make case folding defaults correct for non-Linux Unix systems
When we specify the case folding defaults, we assume the system is case
sensitive if it is Linux and case insensitive otherwise.  While this
gets Windows, Mac, and Linux correct, it produces the wrong defaults for
other Unix systems, such as FreeBSD and NetBSD, which are also case
sensitive.

Since what we want as the default is for Mac and Windows to be case
insensitive, and everything else to be case sensitive, rewrite the
+build comments to that effect.  To ensure that the Go compiler compiles
the case sensitive versions on all Unix systems and that it is not
limited to just Linux by the "_linux" suffix, rename the files
accordingly with more descriptive names.