Ensure that non-ASCII strings match themselves
In the wildmatch code, we escape the strings in a pattern by iterating
over each byte and then appending the result of each iteration of the
loop to a new string.  If the character is not part of an escape
sequence, we append it by casting it to a string.

However, this does not produce the expected results.  When casting an
integral value, such as a byte, to a string, the integral value is
interpreted as a rune.  Consequently, each byte with value greater than
127 in the original string was being turned into a UTF-8 sequence
corresponding to that byte's value in Latin-1.

Because a string which is overencoded does not match the original
string, any attempt to have a non-ASCII string match itself would fail.
Solve this by taking our byte value and first creating a byte slice, and
then casting that to a string.

For the curious, Google Translate reports that the string used in the
tests is Chinese for "hello world".  Using a string that is not in
Latin-1 is preferable to a Latin-1 string because it makes it less
likely that we're getting things right by accident.