Improve matching of directories
Git has two types of patterns: gitignore patterns and gitattributes
patterns.  The former matches directories, and the latter does not.  In
addition, the former treats a single unrooted directory as appearing
anywhere in the path, and it also treats a directory as matching all of
its contents recursively.

Let's improve our support for gitignore patterns by handling these
unrooted directories, which we had not previously handled correctly.  We
use a new type of pattern which works like a doubleStar, but serializes
differently, to match the directory anywhere in the tree.

We also add an option to handle the contents behavior, in which we
recursively match the contents of a directory.  We use a new type,
trailingComponents, which works like doubleStar except for the
serialization.

None of this behavior is enabled in gitattributes mode because none of
that syntax is allowed there.

We add some additional tests and fix the response to a test which was
not correct for our gitignore behavior, but which should not function
correctly.

Co-authored-by: Chris Darroch <chrisd8088@github.com>