Make the contents flag work on patterns starting with slash
If we had a pattern like /foo/, this pattern would not match
foo/filename.txt because we'd end up with three components: a "foo"
component, an empty component, and a trailing components component.  The
empty component would not match, resulting in an incorrect result.

Let's fix this by ensuring we strip out the empty component by slicing
it off.  At the same time, add several new tests for leading slashes,
since we lack adequate testing of these cases.