Match single components properly in contents mode
When we're in contents mode, we can end up with a trailingComponents
token, which may match a single directory in the path, or it may match
the terminal component.  Let's make sure that if it is the terminal
component that we continue to match it by matching the empty string.

Additionally, make sure we handle the directory case by accepting either
one component or two components with a trailing slash as long as the
component can validly be a directory.

Do some cleanup here by mutating dirs instead of using endItem and
removing the need for extra variables in the return statement, while
we're at it.

Co-authored-by: Chris Darroch <chrisd8088@github.com>